refactor: 인터페이스명 변경에 따른 exception 폴더 제거
