+++ src/main/java/com/plzgraduate/myongjigraduatebe/graduation/domain/service/major/MajorManager.java
@@ -7,9 +7,6 @@ import java.util.stream.Collectors;
 import com.plzgraduate.myongjigraduatebe.graduation.domain.model.DetailCategoryResult;
 import com.plzgraduate.myongjigraduatebe.graduation.domain.model.DetailGraduationResult;
 import com.plzgraduate.myongjigraduatebe.graduation.domain.service.GraduationManager;
-import com.plzgraduate.myongjigraduatebe.graduation.domain.service.major.exception.MandatoryMajorSpecialCaseHandler;
-import com.plzgraduate.myongjigraduatebe.graduation.domain.service.major.exception.OptionalMandatoryMajoHandler;
-import com.plzgraduate.myongjigraduatebe.graduation.domain.service.major.exception.ReplaceMandatoryMajorHandler;
 import com.plzgraduate.myongjigraduatebe.lecture.domain.model.Lecture;
 import com.plzgraduate.myongjigraduatebe.lecture.domain.model.MajorLecture;
 import com.plzgraduate.myongjigraduatebe.takenlecture.domain.model.TakenLecture;
